+++ src/Html/DataTableHtml.php
@@ -11,14 +11,15 @@ use Yajra\DataTables\Contracts\DataTableHtmlBuilder;
 abstract class DataTableHtml implements DataTableHtmlBuilder
 {
     /**
-     * @var \Yajra\DataTables\Html\Builder
+     * @var \Yajra\DataTables\Html\Builder|null
      */
-    protected $htmlBuilder;
+    protected ?Builder $htmlBuilder;
 
     /**
      * @return \Yajra\DataTables\Html\Builder
+     * @throws \Exception
      */
-    public static function make()
+    public static function make(): Builder
     {
         if (func_get_args()) {
             return (new static(...func_get_args()))->handle();
@@ -28,25 +29,25 @@ abstract class DataTableHtml implements DataTableHtmlBuilder
     }
 
     /**
-     * @param  string  $name
-     * @param  mixed  $arguments
+     * @param  string  $method
+     * @param  mixed  $parameters
      * @return mixed
      *
      * @throws \Exception
      */
-    public function __call($name, $arguments)
+    public function __call(string $method, $parameters)
     {
-        if (method_exists($this->getHtmlBuilder(), $name)) {
-            return $this->getHtmlBuilder()->{$name}(...$arguments);
+        if (method_exists($this->getHtmlBuilder(), $method)) {
+            return $this->getHtmlBuilder()->{$method}(...$parameters);
         }
 
-        throw new BadMethodCallException("Method {$name} does not exists");
+        throw new BadMethodCallException("Method {$method} does not exists");
     }
 
     /**
      * @return \Yajra\DataTables\Html\Builder
      */
-    protected function getHtmlBuilder()
+    protected function getHtmlBuilder(): Builder
     {
         if ($this->htmlBuilder) {
             return $this->htmlBuilder;
@@ -56,10 +57,10 @@ abstract class DataTableHtml implements DataTableHtmlBuilder
     }
 
     /**
-     * @param  mixed  $builder
+     * @param  \Yajra\DataTables\Html\Builder  $builder
      * @return static
      */
-    public function setHtmlBuilder($builder)
+    public function setHtmlBuilder(Builder $builder): static
     {
         $this->htmlBuilder = $builder;