+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - Added Laravel 9 specific support
 - Bump major version to match with the framework
 - Move `maatwebsite/excel` to suggest. #139
-- [BC] Rename ReadOnly to ReadOnlyField
+- Add phpstan static analysis
+- Add tests